Rename file extension .h -> .hpp
